# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Retail Expansion Intelligence System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List candidate locations
      parameters:
        - name: marketId
          in: query
          schema: { type: ["integer", "null"] }
        - name: state
          in: query
          schema: { type: ["string", "null"] }
        - name: city
          in: query
          schema: { type: ["string", "null"] }
        - name: minScore
          in: query
          schema: { type: ["number", "null"] }
        - name: status
          in: query
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Add a candidate location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      summary: Get location details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationDetail"
        "404":
          description: Not found
    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Update location
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete location
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /locations/{id}/score:
    post:
      operationId: scoreLocation
      tags: [locations]
      summary: Run AI scoring for a location
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Score result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreResult"
  /locations/{id}/financial:
    get:
      operationId: getLocationFinancial
      tags: [locations]
      summary: Get financial simulation for a location
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSimulation"
    patch:
      operationId: updateLocationFinancial
      tags: [locations]
      summary: Update financial inputs for a location
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSimulation"
  /locations/{id}/competitors:
    get:
      operationId: getLocationCompetitors
      tags: [locations]
      summary: Get competitors near a location
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competitor"

  # ─── Markets ────────────────────────────────────────────────
...
components:
  schemas:
...
    Location:
      type: object
      required: [id, name, state, city, latitude, longitude, status, overallScore, recommendation, riskLevel, confidenceLevel, createdAt]
      properties:
        id:           { type: integer }
        name:         { type: string }
        address:      { type: ["string", "null"] }
        state:        { type: string }
        district:     { type: ["string", "null"] }
        city:         { type: string }
        ward:         { type: ["string", "null"] }
        microMarket:  { type: ["string", "null"] }
        latitude:     { type: number }
        longitude:    { type: number }
        status:       { type: string, enum: [candidate, shortlisted, approved, rejected] }
        marketId:     { type: ["integer", "null"] }
        overallScore: { type: number }
        demandScore:  { type: ["number", "null"] }
        competitionScore: { type: ["number", "null"] }
        logisticsScore:   { type: ["number", "null"] }
        financialScore:   { type: ["number", "null"] }
        accessibilityScore: { type: ["number", "null"] }
        growthScore:  { type: ["number", "null"] }
        recommendation: { type: string, enum: [expand, watch, avoid, under_review] }
        riskLevel:    { type: string, enum: [low, medium, high] }
        confidenceLevel: { type: string, enum: [low, medium, high] }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
        updatedAt:    { type: ["string", "null"] }
    LocationDetail:
      allOf:
        - $ref: "#/components/schemas/Location"
        - type: object
          properties:
            demandProfile:
              $ref: "#/components/schemas/DemandProfile"
            financialSimulation:
              $ref: "#/components/schemas/FinancialSimulation"
            scoreBreakdown:
              $ref: "#/components/schemas/ScoreBreakdown"
            nearbyCompetitors:
              type: array
              items:
                $ref: "#/components/schemas/Competitor"
    LocationInput:
      type: object
      required: [name, state, city, latitude, longitude]
      properties:
        name:         { type: string }
        address:      { type: string }
        state:        { type: string }
        district:     { type: string }
        city:         { type: string }
        ward:         { type: string }
        microMarket:  { type: string }
        latitude:     { type: number }
        longitude:    { type: number }
        status:       { type: string }
        marketId:     { type: integer }
        notes:        { type: string }
        population:   { type: number }
        populationDensity: { type: number }
        avgMonthlyIncome: { type: number }
        footfallPerDay: { type: number }
        deliveryDemandScore: { type: number }
        roadAccess:   { type: boolean }
        parkingAvailable: { type: boolean }
        publicTransport: { type: boolean }
        walkabilityScore: { type: number }
        rentalCostPerSqft: { type: number }
    LocationUpdate:
      type: object
      properties:
        name:         { type: string }
        address:      { type: string }
        state:        { type: string }
        district:     { type: string }
        city:         { type: string }
        ward:         { type: string }
        microMarket:  { type: string }
        status:       { type: string }
        notes:        { type: string }
        population:   { type: number }
        populationDensity: { type: number }
        avgMonthlyIncome: { type: number }
        footfallPerDay: { type: number }
        deliveryDemandScore: { type: number }
        roadAccess:   { type: boolean }
        parkingAvailable: { type: boolean }
        rentalCostPerSqft: { type: number }
...
    Competitor:
      type: object
      required: [id, name, type, city, state, createdAt]
      properties:
        id:               { type: integer }
        name:             { type: string }
        brand:            { type: ["string", "null"] }
        type:             { type: string, enum: [milma_booth, juice_shop, tea_shop, bubble_tea, cafe, fresh_juice_chain, ice_cream_chain, convenience_store, mall_food_court, petrol_station, other] }
        city:             { type: string }
        state:            { type: string }
        latitude:         { type: ["number", "null"] }
        longitude:        { type: ["number", "null"] }
        address:          { type: ["string", "null"] }
        priceRange:       { type: ["string", "null"], enum: [budget, mid, premium, null] }
        avgRating:        { type: ["number", "null"] }
        reviewCount:      { type: ["integer", "null"] }
        deliveryAvailable: { type: ["boolean", "null"] }
        operatingHours:   { type: ["string", "null"] }
        strengths:        { type: ["string", "null"] }
        weaknesses:       { type: ["string", "null"] }
        marketShare:      { type: ["number", "null"] }
        notes:            { type: ["string", "null"] }
        createdAt:        { type: string }
...
    DemandProfile:
      type: object
      properties:
        workingProfessionals: { type: ["number", "null"] }
        students:             { type: ["number", "null"] }
        hostelResidents:      { type: ["number", "null"] }
        pgResidents:          { type: ["number", "null"] }
        tourists:             { type: ["number", "null"] }
        families:             { type: ["number", "null"] }
        deliveryUsers:        { type: ["number", "null"] }
        officeEmployees:      { type: ["number", "null"] }
        gymUsers:             { type: ["number", "null"] }
        eveningCrowd:         { type: ["number", "null"] }
        weekendMultiplier:    { type: ["number", "null"] }
        apartmentDensity:     { type: ["number", "null"] }
        estimatedDailyFootfall: { type: ["number", "null"] }
        hasNightlife:         { type: ["boolean", "null"] }
        hasShoppingActivity:  { type: ["boolean", "null"] }
        deliveryPlatforms:    { type: ["string", "null"] }
        peakHours:            { type: ["string", "null"] }
        avgBasketValue:       { type: ["number", "null"] }

    # ─── Financial Simulation ──────────────────────────────────
    FinancialSimulation:
      type: object
      properties:
        capex:                { type: ["number", "null"] }
        monthlyRent:          { type: ["number", "null"] }
        monthlyStaffCost:     { type: ["number", "null"] }
        monthlyUtilities:     { type: ["number", "null"] }
        monthlyInventory:     { type: ["number", "null"] }
        monthlyMarketing:     { type: ["number", "null"] }
        monthlyOpex:          { type: ["number", "null"] }
        monthlyRevenue:       { type: ["number", "null"] }
        grossMarginPct:       { type: ["number", "null"] }
        ebitda:               { type: ["number", "null"] }
        breakEvenMonths:      { type: ["number", "null"] }
        roiPct:               { type: ["number", "null"] }
        irrPct:               { type: ["number", "null"] }
        paybackPeriodMonths:  { type: ["number", "null"] }
        year1Revenue:         { type: ["number", "null"] }
        year3Revenue:         { type: ["number", "null"] }
        year5Revenue:         { type: ["number", "null"] }
        deliveryCommissionPct: { type: ["number", "null"] }
    FinancialInput:
      type: object
      properties:
        capex:                { type: number }
        monthlyRent:          { type: number }
        monthlyStaffCost:     { type: number }
        monthlyUtilities:     { type: number }
        monthlyInventory:     { type: number }
        monthlyMarketing:     { type: number }
        estimatedDailyRevenue: { type: number }
        grossMarginPct:       { type: number }
        deliveryCommissionPct: { type: number }

    # ─── AI Scoring ────────────────────────────────────────────
...
    ScoreResult:
      type: object
      required: [locationId, overallScore, recommendation, riskLevel, confidenceLevel, breakdown]
      properties:
        locationId:          { type: integer }
        overallScore:        { type: number }
        recommendation:      { type: string }
        riskLevel:           { type: string }
        confidenceLevel:     { type: string }
        breakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
    ScoreBreakdown:
      type: object
      properties:
        populationDensityScore:   { type: ["number", "null"] }
        youngPopulationScore:     { type: ["number", "null"] }
        officeDensityScore:       { type: ["number", "null"] }
        studentDensityScore:      { type: ["number", "null"] }
        deliveryDemandScore:      { type: ["number", "null"] }
        commercialActivityScore:  { type: ["number", "null"] }
        competitionGapScore:      { type: ["number", "null"] }
        accessibilityScore:       { type: ["number", "null"] }
        rentalAffordabilityScore: { type: ["number", "null"] }
        tourismScore:             { type: ["number", "null"] }
        nightlifeScore:           { type: ["number", "null"] }
        incomeScore:              { type: ["number", "null"] }
        growthProjectionScore:    { type: ["number", "null"] }

    # ─── Dashboard ─────────────────────────────────────────────
...