# lib/api-zod/src/generated/types/demandProfile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retail Expansion Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export interface DemandProfile {
  /** @nullable */
  workingProfessionals?: number | null;
  /** @nullable */
  students?: number | null;
  /** @nullable */
  hostelResidents?: number | null;
  /** @nullable */
  pgResidents?: number | null;
  /** @nullable */
  tourists?: number | null;
  /** @nullable */
  families?: number | null;
  /** @nullable */
  deliveryUsers?: number | null;
  /** @nullable */
  officeEmployees?: number | null;
  /** @nullable */
  gymUsers?: number | null;
  /** @nullable */
  eveningCrowd?: number | null;
  /** @nullable */
  weekendMultiplier?: number | null;
  /** @nullable */
  apartmentDensity?: number | null;
  /** @nullable */
  estimatedDailyFootfall?: number | null;
  /** @nullable */
  hasNightlife?: boolean | null;
  /** @nullable */
  hasShoppingActivity?: boolean | null;
  /** @nullable */
  deliveryPlatforms?: string | null;
  /** @nullable */
  peakHours?: string | null;
  /** @nullable */
  avgBasketValue?: number | null;
}