# lib/api-zod/src/generated/types/scoreResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retail Expansion Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { ScoreBreakdown } from './scoreBreakdown';

export interface ScoreResult {
  locationId: number;
  overallScore: number;
  recommendation: string;
  riskLevel: string;
  confidenceLevel: string;
  breakdown: ScoreBreakdown;
}